+++ classes/site_backup.php
@@ -16,6 +16,7 @@
 
 namespace tool_vault;
 
+use tool_vault\local\checks\backup_precheck_failed;
 use tool_vault\local\checks\check_base;
 use tool_vault\local\checks\configoverride;
 use tool_vault\local\checks\dbstatus;
@@ -200,11 +201,7 @@ class site_backup extends operation_base {
                 $this->prechecks[$chk->get_name()] = $chk;
                 $this->add_to_log('...OK');
             } else if ($chk) {
-                $a = (object)[
-                    'name' => $classname::get_display_name(),
-                    'message' => $chk->get_status_message().' '.$chk->summary(),
-                ];
-                throw new \moodle_exception('error_backupprecheckfailed', 'tool_vault', '', $a);
+                throw new backup_precheck_failed($chk);
             } else {
                 throw new \moodle_exception('error_unabletorunprecheck', 'tool_vault',
                     '', $classname::get_display_name());