+++ progress.php
@@ -73,19 +73,22 @@ $isoldoperation = $operation &&
 
 if ($operation instanceof \tool_vault\local\models\backup_model) {
     if ($isoldoperation) {
-        $url = tool_vault\local\uiactions\backup_details::url(['id' => $operation->id]);
+        $url = tool_vault\local\uiactions\backup_details::url(['id' => $operation->id])->out();
         echo '<p>' . get_string('backupfinished', 'tool_vault', $url) . '</p>';
     } else {
         $data = (new \tool_vault\output\backup_details($operation, null, true, true))->export_for_template($renderer);
         echo $renderer->render_from_template('tool_vault/backup_details', $data);
     }
 } else if ($operation instanceof \tool_vault\local\models\restore_model) {
+    $url = \tool_vault\local\uiactions\restore_details::url(['id' => $operation->id]);
     if ($isoldoperation) {
-        $url = \tool_vault\local\uiactions\restore_details::url(['id' => $operation->id]);
-        echo '<p>' . get_string('restorefinished', 'tool_vault', $url) . '</p>';
+        echo '<p>' . get_string('restorefinished', 'tool_vault', (string)$url) . '</p>';
     } else {
         $data = (new \tool_vault\output\restore_details($operation))->export_for_template($renderer);
         $data['isprogresspage'] = 1;
+        if ($data['errormessage']) {
+            $data['errordetailsurl'] = $url->out(false);
+        }
         echo $renderer->render_from_template('tool_vault/restore_details', $data);
     }
 } else {